--- src/app/views/CatAnio/Anios.tsx
@@ -1,180 +1,0 @@
-import { GridColDef } from "@mui/x-data-grid";
-import { useEffect, useState } from "react";
-import Swal from "sweetalert2";
-import { Toast } from "../../helpers/Toast";
-import { PERMISO, USUARIORESPONSE } from "../../interfaces/UserInfo";
-import { CatalogosServices } from "../../services/catalogosServices";
-import { getPermisos, getUser } from "../../services/localStorage";
-import MUIXDataGrid from "../MUIXDataGrid";
-import ButtonsAdd from "../componentes/ButtonsAdd";
-import ButtonsDeleted from "../componentes/ButtonsDeleted";
-import ButtonsEdit from "../componentes/ButtonsEdit";
-import TitleComponent from "../componentes/TitleComponent";
-import { AniosModal } from "./AniosModal";
-
-export const Anios = () => {
-  const [openSlider, setOpenSlider] = useState(true);
-  const [modo, setModo] = useState("");
-  const [open, setOpen] = useState(false);
-  const [tipoOperacion, setTipoOperacion] = useState(0);
-  const [vrows, setVrows] = useState({});
-  const [bancos, setBancos] = useState([]);
-  const user: USUARIORESPONSE = JSON.parse(String(getUser()));
-
-  const permisos: PERMISO[] = JSON.parse(String(getPermisos()));
-  const [agregar, setAgregar] = useState<boolean>(true);
-  const [editar, setEditar] = useState<boolean>(true);
-  const [eliminar, setEliminar] = useState<boolean>(true);
-
-  const handleAccion = (v: any) => {
-    if (v.tipo === 1) {
-      setTipoOperacion(2);
-      setModo("Editar Registro");
-      setOpen(true);
-      setVrows(v.data);
-    } else if (v.tipo === 2) {
-      Swal.fire({
-        icon: "info",
-        title: "¿Estás seguro de eliminar este registro?",
-        showDenyButton: true,
-        showCancelButton: false,
-        confirmButtonText: "Confirmar",
-        denyButtonText: `Cancelar`,
-      }).then((result) => {
-        if (result.isConfirmed) {
-          let data = {
-            NUMOPERACION: 3,
-            CHID: v.data.row.id,
-            CHUSER: user.Id,
-          };
-
-          CatalogosServices.aniosindex(data).then((res) => {
-            if (res.SUCCESS) {
-              Toast.fire({
-                icon: "success",
-                title: "¡Registro Eliminado!",
-              });
-              consulta({ NUMOPERACION: 4 });
-            } else {
-              Swal.fire("¡Error!", res.STRMESSAGE, "error");
-            }
-          });
-        } else if (result.isDenied) {
-          Swal.fire("No se realizaron cambios", "", "info");
-        }
-      });
-    }
-  };
-
-  const columns: GridColDef[] = [
-    {
-      field: "id",
-      headerName: "Identificador",
-      width: 150,
-    },
-    {
-      field: "acciones",
-      disableExport: true,
-      headerName: "Acciones",
-      description: "Campo de Acciones",
-      sortable: false,
-      width: 200,
-      renderCell: (v) => {
-        return (
-          <>
-            <ButtonsEdit
-              handleAccion={handleAccion}
-              row={v}
-              show={true}
-            ></ButtonsEdit>
-            <ButtonsDeleted
-              handleAccion={handleAccion}
-              row={v}
-              show={true}
-            ></ButtonsDeleted>
-          </>
-        );
-      },
-    },
-    { field: "FechaCreacion", headerName: "Fecha de Creación", width: 150 },
-    {
-      field: "UltimaActualizacion",
-      headerName: "Última Actualización",
-      width: 150,
-    },
-    { field: "CreadoPor", headerName: "Creado Por", width: 100 },
-    { field: "ModificadoPor", headerName: "Modificado Por", width: 100 },
-    { field: "anio", headerName: "Año", description: "Año", width: 100 },
-  ];
-
-  const handleClose = () => {
-    setOpen(false);
-    consulta({ NUMOPERACION: 4 });
-  };
-
-  const handleOpen = (v: any) => {
-    setTipoOperacion(1);
-    setModo("Agregar Registro");
-    setOpen(true);
-    setVrows("");
-  };
-
-  const handleEdit = (v: any) => {
-    setTipoOperacion(2);
-    setModo("Módificar Registro");
-    setOpen(true);
-    setVrows(v);
-  };
-
-  const consulta = (data: any) => {
-    CatalogosServices.aniosindex(data).then((res) => {
-      if (res.SUCCESS) {
-        // Toast.fire({
-        //   icon: "success",
-        //   title: "¡Consulta Exitosa!",
-        // });
-        setBancos(res.RESPONSE);
-        setOpenSlider(false);
-      } else {
-        setOpenSlider(false);
-        Swal.fire("¡Error!", res.STRMESSAGE, "error");
-      }
-    });
-  };
-
-  useEffect(() => {
-    permisos.map((item: PERMISO) => {
-      if (String(item.ControlInterno) === "ANIOS") {
-        if (String(item.Referencia) === "AGREG") {
-          setAgregar(true);
-        }
-        if (String(item.Referencia) === "ELIM") {
-          setEliminar(true);
-        }
-        if (String(item.Referencia) === "EDIT") {
-          setEditar(true);
-        }
-      }
-    });
-    consulta({ NUMOPERACION: 4 });
-  }, []);
-
-  return (
-    <div style={{ height: 600, width: "100%", padding: "1%" }}>
-      {open ? (
-        <AniosModal
-          open={open}
-          tipo={tipoOperacion}
-          handleClose={handleClose}
-          dt={vrows}
-        />
-      ) : (
-        ""
-      )}
-
-      <TitleComponent title={"Años Fiscales"} show={openSlider} />
-      <ButtonsAdd handleOpen={handleOpen} agregar={true} />
-      <MUIXDataGrid columns={columns} rows={bancos} />
-    </div>
-  );
-};
